+++ packages/cli/src/ceramic-cli-utils.ts
@@ -117,9 +117,12 @@ export class CeramicCliUtils {
     const config = await this._loadDaemonConfig(configFilepath)
 
     // Environment variables override values from config file
-    if (process.env.CERAMIC_INDEXING_DB_URI) config.indexing.db = process.env.CERAMIC_INDEXING_DB_URI
-    if (process.env.CERAMIC_METRICS_EXPORTER_ENABLED) config.metrics.metricsExporterEnabled = Boolean(process.env.CERAMIC_METRICS_EXPORTER_ENABLED)
-    if (process.env.CERAMIC_METRICS_PORT) config.metrics.metricsPort = Number(process.env.CERAMIC_METRICS_PORT)
+    if (process.env.CERAMIC_INDEXING_DB_URI)
+      config.indexing.db = process.env.CERAMIC_INDEXING_DB_URI
+    if (process.env.CERAMIC_METRICS_EXPORTER_ENABLED)
+      config.metrics.metricsExporterEnabled = Boolean(process.env.CERAMIC_METRICS_EXPORTER_ENABLED)
+    if (process.env.CERAMIC_METRICS_PORT)
+      config.metrics.metricsPort = Number(process.env.CERAMIC_METRICS_PORT)
 
     {
       // CLI flags override values from environment variables and config file
@@ -225,7 +228,7 @@ export class CeramicCliUtils {
 
       console.log(doc.id)
       console.log(JSON.stringify(doc.content, null, 2))
-      depreciationNotice()
+      deprecationNotice()
     })
   }
 
@@ -263,7 +266,7 @@ export class CeramicCliUtils {
       await doc.update(parsedContent, metadata)
 
       console.log(JSON.stringify(doc.content, null, 2))
-      depreciationNotice()
+      deprecationNotice()
     })
   }
 
@@ -275,7 +278,6 @@ export class CeramicCliUtils {
     await CeramicCliUtils._runWithCeramicClient(async (ceramic: CeramicApi) => {
       const stream = await TileDocument.load(ceramic, streamRef)
       console.log(JSON.stringify(stream.content, null, 2))
-      depreciationNotice()
     })
   }
 
@@ -287,7 +289,6 @@ export class CeramicCliUtils {
     await CeramicCliUtils._runWithCeramicClient(async (ceramic: CeramicApi) => {
       const stream = await ceramic.loadStream(streamRef)
       console.log(JSON.stringify(StreamUtils.serializeState(stream.state), null, 2))
-      depreciationNotice()
     })
   }
 
@@ -301,7 +302,7 @@ export class CeramicCliUtils {
     await CeramicCliUtils._runWithCeramicClient(async (ceramic: CeramicApi) => {
       const doc = await TileDocument.load(ceramic, id)
       console.log(JSON.stringify(doc.content, null, 2))
-      depreciationNotice()
+      deprecationNotice()
       doc.subscribe(() => {
         console.log('--- document changed ---')
         console.log(JSON.stringify(doc.content, null, 2))
@@ -320,7 +321,6 @@ export class CeramicCliUtils {
       const stream = await ceramic.loadStream(id)
       const commits = stream.allCommitIds.map((v) => v.toString())
       console.log(JSON.stringify(commits, null, 2))
-      depreciationNotice()
     })
   }
 
@@ -391,7 +391,6 @@ export class CeramicCliUtils {
     await CeramicCliUtils._runWithCeramicClient(async (ceramic: CeramicApi) => {
       const result = await ceramic.pin.add(id)
       console.log(JSON.stringify(result, null, 2))
-      depreciationNotice()
     })
   }
 
@@ -405,7 +404,6 @@ export class CeramicCliUtils {
     await CeramicCliUtils._runWithCeramicClient(async (ceramic: CeramicApi) => {
       const result = await ceramic.pin.rm(id)
       console.log(JSON.stringify(result, null, 2))
-      depreciationNotice()
     })
   }
 
@@ -419,10 +417,23 @@ export class CeramicCliUtils {
     await CeramicCliUtils._runWithCeramicClient(async (ceramic: CeramicApi) => {
       const pinnedStreamIds = []
       const iterator = await ceramic.pin.ls(id)
+      let i = 0
+      let truncated = false
       for await (const id of iterator) {
         pinnedStreamIds.push(id)
+        i++
+        if (i > 20) {
+          truncated = true
+          break
+        }
+      }
+      if (truncated) {
+        console.log('Too many results to show them all, printing the first 20:')
       }
       console.log(JSON.stringify(pinnedStreamIds, null, 2))
+      if (truncated) {
+        console.log('... Additional results were not shown')
+      }
     })
   }
 
@@ -487,7 +498,7 @@ export class CeramicCliUtils {
     const cliConfig = await this._loadCliConfig()
 
     console.log(JSON.stringify(cliConfig, null, 2))
-    depreciationNotice()
+    deprecationNotice()
   }
 
   /**
@@ -510,7 +521,7 @@ export class CeramicCliUtils {
 
     console.log(`Ceramic CLI configuration ${variable} set to ${value}`)
     console.log(JSON.stringify(cliConfig))
-    depreciationNotice()
+    deprecationNotice()
   }
 
   /**
@@ -525,7 +536,7 @@ export class CeramicCliUtils {
 
     console.log(`Ceramic CLI configuration ${variable} unset`)
     console.log(JSON.stringify(cliConfig, null, 2))
-    depreciationNotice()
+    deprecationNotice()
   }
 
   /**
@@ -623,7 +634,7 @@ export class CeramicCliUtils {
   }
 }
 
-const depreciationNotice = () => {
+const deprecationNotice = () => {
   console.log(
     `${pc.red(pc.bold('This command has been deprecated.'))}
 Please use the upgraded Glaze CLI instead.