+++ packages/core/src/store/pin-store-factory.ts
@@ -7,7 +7,7 @@ import type { CID } from 'multiformats/cid'
 import path from 'path'
 import os from 'os'
 import { IpfsPinning } from '@ceramicnetwork/pinning-ipfs-backend'
-import { Repository } from '../state-management/repository'
+import { Repository } from '../state-management/repository.js'
 
 export const DEFAULT_STATE_STORE_DIRECTORY = path.join(os.homedir(), '.ceramic', 'statestore')
 const IPFS_GET_TIMEOUT = 60000 // 1 minute