+++ packages/cli/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.5.0](/compare/@ceramicnetwork/cli@2.5.0-rc.1...@ceramicnetwork/cli@2.5.0) (2022-08-08)
+
+**Note:** Version bump only for package @ceramicnetwork/cli
+
+
+
+
+
 # [2.5.0-rc.1](/compare/@ceramicnetwork/cli@2.5.0-rc.0...@ceramicnetwork/cli@2.5.0-rc.1) (2022-07-26)
 
 **Note:** Version bump only for package @ceramicnetwork/cli