+++ packages/stream-model-instance-handler/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.4.0](/compare/@ceramicnetwork/stream-model-instance-handler@0.4.0-rc.2...@ceramicnetwork/stream-model-instance-handler@0.4.0) (2022-08-08)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-model-instance-handler
+
+
+
+
+
 # [0.4.0-rc.2](/compare/@ceramicnetwork/stream-model-instance-handler@0.4.0-rc.1...@ceramicnetwork/stream-model-instance-handler@0.4.0-rc.2) (2022-07-26)
 
 **Note:** Version bump only for package @ceramicnetwork/stream-model-instance-handler